:art: Day 1 & 2 MVP complete, made resize() DRY
Was not very DRY to create a copy of the insert() method to be used on the new_storage in resize, so added a "store" param that defaults to None in insert, and will then be assigned to self.storage. In resize when insert is called, can now pass in new_storage as the store to insert all the copied values over without repeating code.